First cut at high contrast mode using a high contrast main stylesheet and a high-contrast chart stylesheet.
NOT DONE! This system is working best in Firefox. There are still style issues in Google Chrome, and there likely will be others in IE.

Moreover, there is likely much fine-tuning to be done of the colors, font-sizes, etc.  I committed to get the system in place so Shen and Ivan can contribute.