Meta IE compatability+ add The to map takeaway
Added the IE-edge metatag in the head, to improve the chances the page will work in old versions of Internet Explorer.

Our web team had to add this last time because: We put the tag <meta http-equiv="x-ua-compatible" content="IE=edge" > into the head of our pages to force IE to use standards mode, regardless of what the browser settings are. Otherwise IE 9 might be in a compatibility mode, which means the Javascript doesn't behave correctly. I think the Department default setting is to have compatibility mode on for internal sites, which is why the code that was fine on github behaved badly on an ed.gov server.